Fix Dozer bug with saving LocalDateTime
The old version of Dozer used by the project has a bug that prohibits it
from mapping LocalDateTime. Thus, I would not be able to persist
SingleLoan, which uses LocalDateTime.

https://github.com/DozerMapper/dozer/issues/172

The bug has been fixed in 6.4.0.
Java 8 date and time support (#685)

The latest version is 6.5.0, which also fixes a related bug.
feat: Add converter for java.time.Instant (#731) (#742)

Therefore, I have upgraded from 5.5.1 to 6.5.0 which fixed the problem.
There is also a new test for loanFacade, which makes use of this fix.

There are also some important bugfixes and refactorings in this commit.