Initial AngularJS application
For user management. It is supposed to work with HATEOAS REST API so
either we enable HATEOAS somehow or edit AngularJS source to use our api
format. Also, we need to enable /pa165 context path to access it at
localhost:8080/pa165/users.html as curently it is at the index that
apperars after sign in.