Implement new datepicker in create loan
Remove the obsolete (and broken) datepicker. Use 'AngularJS DateTime
Picker Without JQuery and Bootstrap' instead. It works fine.

The previous datepicker was broken, because it would not fill the input
filed after selecting the date. The value would appear visually inside
the field, but it would not be sent to the AngularJS model
(loan.registeredAt) for some reason. Interestingly, if the user adds any
character manually into the selected date, then the value is sent to
loan.registeredAt and ng-reguired is satisfied. It isa  bug of some
sort.