Open books/users in a new tab
In order to create a loan a user must first find the id of the
appropriate book/loan. They use the "Find" link to do so. However, the
content of the input tabs would get lost (or misbehave) when leaving the
Create loan page. Hence, we open the books/users lists in a new tab.